refactor(architecture): Split monolithic users module into bounded contexts
Split users module into three separate bounded contexts:
- auth/ - authentication logic (login, logout, refresh, forgot password)
- users/ - user management (registration, email confirmation, password reset)
- shared/ - common infrastructure (UoW, services, base exceptions)

This separation provides:
- Clear boundaries between authentication and user management
- Reusable infrastructure components
- Better alignment with DDD principles"
