+++ src/api/users/exceptions/error_messages.py
@@ -1,21 +1,24 @@
 from fastapi import status
 
 from src.core.users.application.exceptions.exception_classes import ServiceUnavailableError, InternalServerError
-from src.core.users.domain.exceptions.exception_classes import UserAlreadyExistsError, MissingRequiredFieldError, \
-    InvalidInputError, InvalidCredentialsError, EmailNotConfirmedError, InvalidEmailConfirmationCodeError, \
-    ConfirmationCodeExpiredError, TokenCryptographyError, TokenStateError
 
-ERROR_MESSAGES = {
+from src.core.users.domain.exceptions.exception_classes import (
+    UserAlreadyExistsError,
+    MissingRequiredFieldError,
+    InvalidInputError,
+    EmailNotConfirmedError,
+    InvalidEmailConfirmationCodeError,
+    ConfirmationCodeExpiredError
+)
+
+
+USER_ERROR_MESSAGES = {
     # DOMAIN
     "user_already_exists": "Такой {verbose_name} уже зарегистрирован",
-    "invalid_login_credentials": "Неправильный email или пароль",
 
     "email_not_confirmed": "Ваш email адрес не подтвержден. Пожалуйста, проверьте почту для завершения регистрации",
     "invalid_confirmation_code": "Неправильная ссылка подтверждения электронной почты. Запросите новое письмо.",
     "confirmation_code_expired": "Ссылка для подтверждения электронной почты истекла. Запросите новую.",
-
-    "token_cryptography_error": "Сессия недействительна. Пожалуйста, войдите заново",
-    "token_state_error": "Сессия недействительна. Пожалуйста, войдите заново",
 
     # DOMAIN VALUE OBJECTS
     "missing_required_field": "{verbose_name} обязательное поле",
@@ -30,6 +33,7 @@ ERROR_MESSAGES = {
     "internal_server_error": "Технические неполадки. Пожалуйста, попробуйте позднее",
 }
 
+
 PYDANTIC_ERROR_MESSAGES = {
     "missing": "{verbose_name} обязательно для заполнения",
     "string_type": "{verbose_name} должно быть строкой",
@@ -37,36 +41,14 @@ PYDANTIC_ERROR_MESSAGES = {
 }
 
 
-def get_error_message(code: str, verbose_name: str) -> str:
-    template = ERROR_MESSAGES.get(code, "Произошла ошибка")
-
-    if verbose_name and "{verbose_name}" in template:
-        return template.format(verbose_name=verbose_name)
-
-    return template
-
-
-def get_pydantic_error_message(code: str, verbose_name: str) -> str:
-    template = PYDANTIC_ERROR_MESSAGES.get(code, "Произошла ошибка")
-
-    if verbose_name and "{verbose_name}" in template:
-        return template.format(verbose_name=verbose_name)
-
-    return template
-
-
-HTTP_STATUS_MAP = {
+USER_HTTP_STATUS_MAP = {
     UserAlreadyExistsError: status.HTTP_409_CONFLICT,
     MissingRequiredFieldError: status.HTTP_422_UNPROCESSABLE_ENTITY,
     InvalidInputError: status.HTTP_422_UNPROCESSABLE_ENTITY,
 
-    InvalidCredentialsError: status.HTTP_401_UNAUTHORIZED,
     EmailNotConfirmedError: status.HTTP_401_UNAUTHORIZED,
     InvalidEmailConfirmationCodeError: status.HTTP_400_BAD_REQUEST,
     ConfirmationCodeExpiredError: status.HTTP_400_BAD_REQUEST,
-
-    TokenCryptographyError: status.HTTP_401_UNAUTHORIZED,
-    TokenStateError: status.HTTP_401_UNAUTHORIZED,
 
     ServiceUnavailableError: status.HTTP_503_SERVICE_UNAVAILABLE,
     InternalServerError: status.HTTP_500_INTERNAL_SERVER_ERROR
