+++ src/core/auth/application/abstract_auth_uow.py
@@ -1,0 +1,10 @@
+from abc import abstractmethod
+
+from src.core.shared.application.abstract_uow import AbstractUnitOfWork
+
+
+class AbstractAuthUnitOfWork(AbstractUnitOfWork):
+    @property
+    @abstractmethod
+    def token(self):
+        raise NotImplementedError
