+++ src/core/auth/application/exceptions/exception_classes.py
