+++ src/core/auth/application/usecases/refresh_token.py
@@ -1,29 +1,30 @@
 import uuid
 from typing import Tuple
 
-from src.api.users.user_dto import LoginResponseDTO, TokenDataDTO, RefreshTokenDTO
-from src.core.users.application.exceptions.exception_classes import ServiceUnavailableError, InternalServerError
-from src.core.users.application.uow import AbstractUserUnitOfWork
-from src.core.users.domain.enums import TokenTypeEnum
-from src.core.users.domain.exceptions.exception_classes import TokenStateError, TokenCryptographyError
+from src.api.users.user_dto import TokenDataDTO
+from src.api.auth.auth_dto import LoginResponseDTO, RefreshTokenDTO
+from src.core.auth.application.exceptions.exception_classes import ServiceUnavailableError, InternalServerError
+from src.core.auth.application.abstract_auth_uow import AbstractAuthUnitOfWork
+from src.core.auth.domain.enums import TokenTypeEnum
+from src.core.auth.domain.exceptions.exception_classes import TokenStateError, TokenCryptographyError
 
-from src.core.users.infrastructure.exceptions.exception_classes import (
+from src.core.shared.infrastructure.exceptions.exception_classes import DatabaseConnectionError, UnitOfWorkError
+
+from src.core.auth.infrastructure.exceptions.exception_classes import (
     TokenExpiredError,
     InvalidTokenError,
-    UnitOfWorkError,
-    DatabaseConnectionError,
     RepositoryError,
     TokenGeneratorService
 )
 
-from src.core.users.infrastructure.factories import AuthTokenFactory
-from src.core.users.infrastructure.services.pyjwt_token import PyJWTTokenService
+from src.core.auth.infrastructure.factories import AuthTokenFactory
+from src.core.auth.infrastructure.services.pyjwt_token import PyJWTTokenService
 
 
 class RefreshTokenUseCase:
     def __init__(
         self,
-        unit_of_work: AbstractUserUnitOfWork,
+        unit_of_work: AbstractAuthUnitOfWork,
         token_service: PyJWTTokenService
     ):
         self.unit_of_work = unit_of_work
