+++ src/core/auth/domain/exceptions/exception_classes.py
@@ -1,0 +1,26 @@
+from src.core.shared.application.exceptions.base_exception import BaseApplicationError
+
+
+class DomainLayerError(BaseApplicationError):
+    pass
+
+
+# ==========================================================
+# TOKENS ERRORS
+# ==========================================================
+
+
+class TokenAlreadyRevokedError(DomainLayerError):
+    pass
+
+
+class TokenCryptographyError(DomainLayerError):
+    pass
+
+
+class TokenStateError(DomainLayerError):
+    pass
+
+
+class InvalidCredentialsError(DomainLayerError):
+    pass
