+++ src/core/auth/infrastructure/auth_unit_of_work.py
@@ -1,21 +1,17 @@
-from typing import Callable, Optional
+from typing import Callable
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from src.core.auth.infrastructure.auth_repository import AuthTokenRepository
-from src.core.shared.infrastructure.sql_alchemy_uow import SQLAlchemyUnitOfWork
-from src.core.users.infrastructure.user_repository import UserRepository
+from src.core.users.infrastructure.user_unit_of_work import SQLAlchemyUnitOfWork
 
 
-class UserUnitOfWork(SQLAlchemyUnitOfWork):
+class AuthUnitOfWork(SQLAlchemyUnitOfWork):
     def __init__(self, session_factory: Callable[[], AsyncSession]):
         super().__init__(session_factory=session_factory)
 
     @property
-    def user(self):
+    def token(self):
         if self._session is None:
             raise RuntimeError("UoW not initialized!")
-        return UserRepository(self._session)
-
-
-
+        return AuthTokenRepository(self._session)
