+++ src/core/shared/application/exceptions/base_exception.py
