+++ src/core/users/application/abstract_user_uow.py
@@ -1,0 +1,13 @@
+from abc import abstractmethod
+
+from src.core.shared.application.abstract_uow import AbstractUnitOfWork
+
+
+class AbstractUserUnitOfWork(AbstractUnitOfWork):
+    @property
+    @abstractmethod
+    def user(self):
+        raise NotImplementedError
+
+
+
