+++ src/core/users/domain/entities.py
@@ -3,8 +3,8 @@ from dataclasses import dataclass
 from datetime import datetime, timezone
 from typing import Union
 
-from src.core.users.domain.enums import UserRoleEnum, BusinessTypeEnum, TokenTypeEnum
-from src.core.users.domain.exceptions.exception_classes import EmailAlreadyConfirmedError, TokenAlreadyRevokedError
+from src.core.users.domain.enums import UserRoleEnum, BusinessTypeEnum
+from src.core.users.domain.exceptions.exception_classes import EmailAlreadyConfirmedError
 from src.core.users.domain.value_objects import Fullname, Email, Phone, OrganizationFullname, IIN, BIN, HashedPassword
 
 
@@ -49,20 +49,4 @@ class UserAuthEntity:
     password: HashedPassword
 
 
-@dataclass
-class AuthTokenAggregate:
-    id: uuid.UUID
-    user_id: uuid.UUID
-    token_type: TokenTypeEnum
-    token_value: str
-    is_revoked: bool
-    expires_at: datetime
 
-    def is_expired(self):
-        current_time = datetime.now(timezone.utc)
-        return current_time > self.expires_at
-
-    def revoke_token(self):
-        if self.is_revoked:
-            raise TokenAlreadyRevokedError(code="token_already_revoked")
-        self.is_revoked = True
