+++ src/core/users/infrastructure/factories.py
@@ -1,10 +1,8 @@
 import uuid
-from datetime import datetime
 
 from src.api.users.user_dto import UserDTO, IndividualUserDTO, BusinessUserDTO, UserAuthDTO
-from src.core.users.domain.entities import IndividualUserEntity, BusinessUserEntity, UserAggregate, UserAuthEntity, \
-    AuthTokenAggregate
-from src.core.users.domain.enums import UserRoleEnum, TokenTypeEnum
+from src.core.users.domain.entities import IndividualUserEntity, BusinessUserEntity, UserAggregate, UserAuthEntity
+from src.core.users.domain.enums import UserRoleEnum
 from src.core.users.domain.value_objects import Fullname, Email, Phone, OrganizationFullname, IIN, BIN, HashedPassword
 
 
@@ -64,59 +62,3 @@ class UserAuthFactory:
             password=HashedPassword.from_raw(auth_dto.password)
         )
 
-
-class AuthTokenFactory:
-    @staticmethod
-    def create_refresh_token(
-        user_id: uuid.UUID,
-        token_value: str,
-        expires_at: datetime
-    ):
-        return AuthTokenFactory._create(
-            user_id=user_id,
-            token_type=TokenTypeEnum.REFRESH_TOKEN,
-            token_value=token_value,
-            expires_at=expires_at
-        )
-
-    @staticmethod
-    def create_email_token(
-        user_id: uuid.UUID,
-        token_value: str,
-        expires_at: datetime
-    ):
-        return AuthTokenFactory._create(
-            user_id=user_id,
-            token_type=TokenTypeEnum.EMAIL_CONFIRMATION_TOKEN,
-            token_value=token_value,
-            expires_at=expires_at
-        )
-
-    @staticmethod
-    def create_password_reset_token(
-        user_id: uuid.UUID,
-        token_value: str,
-        expires_at: datetime
-    ):
-        return AuthTokenFactory._create(
-            user_id=user_id,
-            token_type=TokenTypeEnum.PASSWORD_RESET_TOKEN,
-            token_value=token_value,
-            expires_at=expires_at
-        )
-
-    @staticmethod
-    def _create(
-        user_id: uuid.UUID,
-        token_type: TokenTypeEnum,
-        token_value: str,
-        expires_at: datetime
-    ):
-        return AuthTokenAggregate(
-            id=uuid.uuid4(),
-            user_id=user_id,
-            token_type=token_type,
-            token_value=token_value,
-            is_revoked=False,
-            expires_at=expires_at
-        )
