+++ src/main.py
@@ -1,10 +1,10 @@
-import uvicorn
 from fastapi import FastAPI
 from fastapi.exceptions import RequestValidationError
 
-from src.api.users.exceptions.exception_handlers import application_exception_handler, pydantic_exception_handler
+from src.api.shared.exceptions.exception_handlers import application_exception_handler, pydantic_exception_handler
+from src.api.auth.auth_routers import auth_router
 from src.api.users.user_routers import users_router
-from src.base_exception import BaseApplicationError
+from src.core.shared.application.exceptions.base_exception import BaseApplicationError
 from src.configuration.conf.settings import DatabaseConnectionSettings
 from src.configuration.dependencies.depends import DependencyContainer
 
@@ -21,6 +21,8 @@ def create_fastapi_app() -> FastAPI:
     app.add_exception_handler(RequestValidationError, pydantic_exception_handler)
 
     app.include_router(users_router)
+    app.include_router(auth_router)
+
     return app
